+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 **Features**
 
 - Registry: Update with new OIDs from OpenWB.
+- Tool `read_pcap.py` now makes an attempt to decode frames that are complete but have an incorrect checksum to try to
+  give a better insight into what's going on.
+- Tool `read_pcap.py` prints the time stamp encoded in the dump with each packet.
+- `ReceiveFrame`: Add a flag to allow decoding the content of a frame even if the CRC checksum does not match. This is
+  intended as a debug measure and not to be used in normal operation.
 
 **Documentation**
 
@@ -24,6 +29,10 @@ All notable changes to this project will be documented in this file.
   content would not make sense. This change allows users to filter these our, e.g. when printing their content.
 - Simulator: If multiple requests were sent in the same TCP paket, the simulator returned the answer for the first
   frame that it got for all of the requests in the buffer.
+- Tool `read_pcap.py` now drops a frame if it ran over the segment boundary (next tcp packet) if the new segment looks
+  like it starts with a new frame (`0x002b`). This way invalid frames with very high length fields are caught earlier,
+  only losing the rest of the segment instead of consuming potentially hundrets of frames only to error out on
+  CRC-check.
 
 ## Release 0.0.2 - 2021-02-17